# ENH: Add ``__slots__`` to private (sub-)classes in ``numpy.lib._index_tricks_impl``