--- a/numpy/ma/extras.py
+++ b/numpy/ma/extras.py
@@ -1014,7 +1014,7 @@ def compress_rows(a):
     ...                                                   [0, 0, 0]])
     >>> np.ma.compress_rows(a)
     array([[6, 7, 8]])
-    
+
     """
     a = asarray(a)
     if a.ndim != 2:
@@ -1387,7 +1387,7 @@ def setxor1d(ar1, ar2, assume_unique=False):
     >>> ar2 = np.ma.array([2, 3, 5, 7, 5])
     >>> np.ma.setxor1d(ar1, ar2)
     masked_array(data=[1, 4, 5, 7],
-                 mask=False,    
+                 mask=False,
            fill_value=999999)
 
     """
@@ -1569,8 +1569,8 @@ def _covhelper(x, y=None, rowvar=True, allow_masked=True):
         tup = (None, slice(None))
     #
     if y is None:
-        # Check if we can guarantee that the integers in the (N - ddof) 
-        # normalisation can be accurately represented with single-precision 
+        # Check if we can guarantee that the integers in the (N - ddof)
+        # normalisation can be accurately represented with single-precision
         # before computing the dot product.
         if x.shape[0] > 2 ** 24 or x.shape[1] > 2 ** 24:
             xnm_dtype = np.float64
@@ -1591,8 +1591,8 @@ def _covhelper(x, y=None, rowvar=True, allow_masked=True):
                     x._sharedmask = False
                     y._sharedmask = False
         x = ma.concatenate((x, y), axis)
-        # Check if we can guarantee that the integers in the (N - ddof) 
-        # normalisation can be accurately represented with single-precision 
+        # Check if we can guarantee that the integers in the (N - ddof)
+        # normalisation can be accurately represented with single-precision
         # before computing the dot product.
         if x.shape[0] > 2 ** 24 or x.shape[1] > 2 ** 24:
             xnm_dtype = np.float64
@@ -1673,7 +1673,7 @@ def cov(x, y=None, rowvar=True, bias=False, allow_masked=True, ddof=None):
           [ True,  True,  True,  True]],
     fill_value=1e+20,
     dtype=float64)
-    
+
     """
     # Check inputs
     if ddof is not None and ddof != int(ddof):
@@ -1791,6 +1791,8 @@ class MAxisConcatenator(AxisConcatenator):
     mr_class
 
     """
+    __slots__ = ()
+
     concatenate = staticmethod(concatenate)
 
     @classmethod
@@ -1828,6 +1830,8 @@ class mr_class(MAxisConcatenator):
            fill_value=999999)
 
     """
+    __slots__ = ()
+
     def __init__(self):
         MAxisConcatenator.__init__(self, 0)