This adds ``__slots__`` to the (private) classes in ``numpy.lib._index_tricks_impl`` and their subclasses, improving the performance of:

- ``numpy.mgrid``
- ``numpy.ogrid``
- ``numpy.r_``
- ``numpy.c_``
- ``numpy.index_exp``
- ``numpy.s_``
- ``numpy.ma.mr_``

> [!NOTE]
> I initially wanted to add ``__slots__`` to ``ndenumerate`` and ``ndindex`` as well, but decided not to, as there's a (small) chance that it will be backwards compatible.: 
> Because they are part of the public API, and not marked as ``@final``, it's possible that > they're used as base-class for user-defined types. And while that, on its own, is not a problem here, it could become one if such a user-defined type has more than one base class.
> 
> But if this isn't a problem, I'd be happy to add ``__slots__`` to these classes, too :)