@eendebakpt 2024-07-20
@jorenham Do you have some benchmarks for the methods affected by these changes? I suspect this will mostly impact operations involving small arrays