@jorenham 2024-07-21
@eendebakpt A quick `%timeit` comparison on Python 3.12 gave the following results:

|                         | numpy/numpy@main       | jorenham/numpy@index_tricks-__slots__ | Speedup |
|-------------------------|------------------------|---------------------------------------|-------------|
| `np.mgrid[0:2]`           | 3.74 μs ± 10.1 ns      | 3.18 μs ± 7.03 ns                     | + 17.6 %    |
| `np.mgrid[0:1000]`        | 4.78 μs ± 7.69 ns      | 4.12 μs ± 10.4 ns                     | + 16.0 %    |
| `np.mgrid[0:2, 0:2]`      | 23.1 μs ± 49.3 ns      | 20.4 μs ± 43.3 ns                     | + 14.9 %    |
| `np.r_[0, 1]`             | 6.65 μs ± 9.88 ns      | 5.61 μs ± 8.72 ns                     | + 18.5 %    |
| `np.r_[0, 1, 2, 3]`       | 8.34 μs ± 8.4 ns       | 7.15 μs ± 15 ns                       | + 16.6 %    |
| `np.r_[[[0]], [[1]]]`     | 20.1 μs ± 31 ns        | 17.5 μs ± 30.3 ns                     | + 14.9 %    |
| `np.s_[2::2]`             | 95.4 ns ± 0.0396 ns    | 45.7 ns ± 0.0559 ns                   | + 108 %     |

The remaining `ogrid`, `c_`/`ma.mr_`, and `index_exp` should have similar results as `mgrid`, `r_`, and `s_`, respectively.